Tighten text-field's `preset` prop's type
Fixes https://github.com/infinitered/ignite-bowser/issues/388
